ml/requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
dill==0.2.7.1
future==0.16.0
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-pubsub==0.28.3
google-cloud-storage==1.4.0
google-gax==0.15.15
google-resumable-media==0.2.3
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.6.0
idna==2.6
numpy==1.13.1
ply==3.8
protobuf==3.4.0
psutil==5.3.1
pyasn1==0.3.6
pyasn1-modules==0.1.4
requests==2.18.4
rsa==3.4.2
scikit-surprise==1.0.3
six==1.11.0
urllib3==1.22
